requirements.txt
# ============================================
# AI Medical Data Platform - Dependencies
# ============================================

# Local LLMs via Ollama (RECOMMENDED for medical data privacy)
ollama>=0.1.6

# Cloud LLM APIs (optional alternatives)
anthropic>=0.39.0
openai>=1.12.0
google-generativeai>=0.5.0

# Neo4j Knowledge Graph
neo4j>=5.16.0

# API Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3

# Analytics Dashboard
dash>=2.14.2
plotly>=5.18.0
pandas>=2.1.4

# Configuration & HTTP
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.25.0

# ============================================
# OPEN-SOURCE MODEL SUPPORT (Optional)
# ============================================
# Uncomment below to enable HuggingFace models locally
# WARNING: Adds ~2GB download, much slower build
#
# torch==2.1.2
# transformers==4.37.2
# accelerate==0.26.1
# bitsandbytes==0.42.0
# sentencepiece==0.1.99
